+++ src/main/java/requious/recipe/RequirementSelection.java
@@ -48,7 +48,7 @@ public class RequirementSelection extends RequirementBase {
 
     @Override
     public ConsumptionResult createResult() {
-        return new ConsumptionResult.Integer(this, 0);
+        return new ConsumptionResult.Long(this, 0L);
     }
 
     @Override
